Добавила ValidateCrudRequester2 на тесты, где есть запрос на просмотр трансфера и добавила endpoint LOOK_TRANSFER в класс Endpoint

+++ src/main/java/iteration_2/requests/skelethon/interfaces/CrudEndpointInterface.java
@@ -4,7 +4,7 @@ import iteration_2.models_body_JSON.BaseModel;
 
 public interface CrudEndpointInterface {
 	Object post(BaseModel baseModel);
-	Object get(long id);
+	Object get(int id);
 	Object update(long id, BaseModel baseModel);
 	Object delete(long id);
 }
